Refactor
