+++ lib/ui/collections/album/list_item.dart
