+++ lib/ui/viewer/gallery/archive_page.dart
@@ -9,7 +9,7 @@ import 'package:photos/models/gallery_type.dart';
 import "package:photos/models/metadata/common_keys.dart";
 import 'package:photos/models/selected_files.dart';
 import 'package:photos/services/collections_service.dart';
-import "package:photos/ui/components/album_horizontal_list_widget.dart";
+import "package:photos/ui/collections/album/horizontal_list.dart";
 import 'package:photos/ui/viewer/actions/file_selection_overlay_bar.dart';
 import "package:photos/ui/viewer/gallery/empty_state.dart";
 import 'package:photos/ui/viewer/gallery/gallery.dart';
@@ -68,7 +68,7 @@ class ArchivePage extends StatelessWidget {
       emptyState: EmptyState(
         text: S.of(context).youDontHaveAnyArchivedItems,
       ),
-      header: AlbumHorizontalListWidget(
+      header: AlbumHorizontalList(
         CollectionsService.instance.getArchivedCollectionWithThumb,
       ),
     );
